+++ src/datasources/results/components/query-builders/query-steps/StepsQueryBuilder.tsx
@@ -17,7 +17,7 @@ import { StepsQueryBuilderFields, StepsQueryBuilderStaticFields } from 'datasour
 type StepsQueryBuilderProps = QueryBuilderProps &
   React.HTMLAttributes<Element> & {
     filter?: string;
-    workspaces: Workspace[];
+    workspaces: Workspace[] | null;
     stepStatus: string[];
     stepsPath: string[];
     globalVariableOptions: QueryBuilderOption[];
@@ -45,6 +45,10 @@ export const StepsQueryBuilder: React.FC<StepsQueryBuilderProps> = ({
   }, [filter]);
 
   const workspaceField = useMemo(() => {
+    if (!workspaces) {
+      return null;
+    }
+    
     const workspaceField = StepsQueryBuilderFields.WORKSPACE;
     return {
       ...workspaceField,
@@ -103,6 +107,10 @@ export const StepsQueryBuilder: React.FC<StepsQueryBuilderProps> = ({
   }, [stepsPath]);
 
   useEffect(() => {
+    if(!workspaceField) {
+      return;
+    }
+
     const updatedFields = [
       stepsPathField,
       updatedAtField,
@@ -153,6 +161,8 @@ export const StepsQueryBuilder: React.FC<StepsQueryBuilderProps> = ({
       QueryBuilderOperations.LIST_DOES_NOT_EQUAL,
       QueryBuilderOperations.LIST_CONTAINS,
       QueryBuilderOperations.LIST_DOES_NOT_CONTAIN,
+      QueryBuilderOperations.DATE_TIME_IS_AFTER,
+      QueryBuilderOperations.DATE_TIME_IS_BEFORE
     ].map(operation => {
       return {
         ...operation,