+++ src/datasources/results/components/variable-editors/ResultsVariableQueryEditor.test.tsx
@@ -2,10 +2,12 @@ import { fireEvent, screen, waitFor } from '@testing-library/react';
 import { ResultsVariableQueryEditor } from './ResultsVariableQueryEditor';
 import { setupRenderer } from 'test/fixtures';
 import { ResultsDataSource } from 'datasources/results/ResultsDataSource';
-import { ResultsQuery } from 'datasources/results/types/types';
+import { QueryType, ResultsQuery } from 'datasources/results/types/types';
 import { Workspace } from 'core/types';
 import { QueryResultsDataSource } from 'datasources/results/query-handlers/query-results/QueryResultsDataSource';
 import { ResultsVariableProperties } from 'datasources/results/types/QueryResults.types';
+import { QueryStepsDataSource } from 'datasources/results/query-handlers/query-steps/QueryStepsDataSource';
+import userEvent from '@testing-library/user-event';
 
 const fakeWorkspaces: Workspace[] = [
   {
@@ -31,33 +33,161 @@ class FakeQueryResultsSource extends QueryResultsDataSource {
   }
 }
 
+class FakeQueryStepsDataSource extends QueryStepsDataSource {
+  getWorkspaces(): Promise<Workspace[]> {
+    return Promise.resolve(fakeWorkspaces);
+  }
+  getPartNumbers(): Promise<void> {
+    return Promise.resolve();
+  }
+}
+
 class FakeResultsDataSource extends ResultsDataSource {
   get queryResultsDataSource() {
     return new FakeQueryResultsSource(this.instanceSettings, this.backendSrv, this.templateSrv);
+  }
+
+  get queryStepsDataSource() {
+    return new FakeQueryStepsDataSource(this.instanceSettings, this.backendSrv, this.templateSrv);
   }
 }
 
 const renderEditor = setupRenderer(ResultsVariableQueryEditor, FakeResultsDataSource, () => {});
 let propertiesSelect: HTMLElement;
 let queryBy: HTMLElement;
+let queryByResults: HTMLElement;
+let queryBySteps: HTMLElement;
 
-it('should render properties select and results query builder progressively', async () => {
-  renderEditor({ refId: '', properties: '', queryBy: '' } as unknown as ResultsQuery);
-  propertiesSelect = screen.getAllByRole('combobox')[0];
+describe('Results Query Type', () => {
+  beforeEach(() => {
+    renderEditor({ refId: '', queryType: QueryType.Results, properties: ResultsVariableProperties[0].value, queryBy: '', resultsTake: 1000 } as unknown as ResultsQuery);
+  });
 
-  expect(propertiesSelect).toBeInTheDocument();
+  it('should render query type radio buttons', () => {
+    const radioButtons = screen.getAllByRole('radio');
 
-  //simulate user selecting a property
-  fireEvent.keyDown(propertiesSelect, { key: 'ArrowDown' });
-  const option = await screen.findByText(ResultsVariableProperties[0].label);
-  fireEvent.click(option);
+    expect(radioButtons.length).toBe(2);
+  });
 
-  queryBy = screen.getByText('Query by results properties');
-  expect(queryBy).toBeInTheDocument();
+  it('should select Results query type by default', () => {
+    expect(screen.getByRole('radio', { name: QueryType.Results })).toBeInTheDocument();
+    expect(screen.getByRole('radio', { name: QueryType.Results })).toBeChecked();
+    expect(screen.getByRole('radio', { name: QueryType.Steps })).toBeInTheDocument();
+    expect(screen.getByRole('radio', { name: QueryType.Steps })).not.toBeChecked();
+  });
 
-  await waitFor(() => expect(screen.getAllByText('Property').length).toBe(1));
-  await waitFor(() => expect(screen.getAllByText('Operator').length).toBe(1));
-  await waitFor(() => expect(screen.getAllByText('Value').length).toBe(1));
+  it('should render properties select and results query builder progressively', async () => {
+    propertiesSelect = screen.getAllByRole('combobox')[0];
+
+    expect(propertiesSelect).toBeInTheDocument();
+
+    //simulate user selecting a property
+    fireEvent.keyDown(propertiesSelect, { key: 'ArrowDown' });
+    fireEvent.click(propertiesSelect);
+
+    queryBy = screen.getByText('Query by results properties');
+    expect(queryBy).toBeInTheDocument();
+
+    await waitFor(() => expect(screen.getAllByText('Property').length).toBe(1));
+    await waitFor(() => expect(screen.getAllByText('Operator').length).toBe(1));
+    await waitFor(() => expect(screen.getAllByText('Value').length).toBe(1));
+  });
+
+  describe('Take input field', () => {
+    it('should render take input field with 1000 as value by default', () => {
+      const takeInput = screen.getByPlaceholderText('Enter record count');
+      expect(takeInput).toBeInTheDocument();
+      expect(takeInput).toHaveValue(1000);
+    });
+
+    it('should only allows numbers in Take field', async () => {
+      const takeInput = screen.getByPlaceholderText('Enter record count');
+
+      // User tries to enter a non-numeric value
+      await userEvent.clear(takeInput);
+      await userEvent.type(takeInput, 'abc');
+      await waitFor(() => {
+        expect(takeInput).toHaveValue(null);
+      });
+
+      // User enters a valid numeric value
+      await userEvent.clear(takeInput);
+      await userEvent.type(takeInput, '500');
+      await waitFor(() => {
+        expect(takeInput).toHaveValue(500);
+      });
+    });
+  });
+});
+
+describe('Steps Query Type', () => {
+  it('should render steps wrapper query builder', async () => {
+    renderEditor({
+      refId: '',
+      queryType: QueryType.Steps,
+      queryByResults: 'resultsQuery',
+      queryBySteps: '',
+    } as unknown as ResultsQuery);
+
+    queryByResults = screen.getByText('Query by results properties');
+    queryBySteps = screen.getByText('Query by steps properties');
+
+    expect(queryByResults).toBeInTheDocument();
+    expect(queryBySteps).toBeInTheDocument();
+  });
+
+  describe('Take input field', () => {
+    it('should render take input field with 1000 as value by default', () => {
+      renderEditor({
+        refId: '',
+        queryType: QueryType.Steps,
+        queryByResults: 'resultsQuery',
+        queryBySteps: '',
+      } as unknown as ResultsQuery);
+
+      const takeInput = screen.getByPlaceholderText('Enter record count');
+      expect(takeInput).toBeInTheDocument();
+      expect(takeInput).toHaveValue(1000);
+    });
+
+    it('should render with existing stepsTake when take is already set', () => {
+      renderEditor({
+        refId: '',
+        queryType: QueryType.Steps,
+        queryByResults: 'resultsQuery',
+        queryBySteps: '',
+        stepsTake: 2000,
+      } as unknown as ResultsQuery);
+
+      const takeInput = screen.getByPlaceholderText('Enter record count');
+      expect(takeInput).toHaveValue(2000);
+    });
+
+    it('should only allows numbers in Take field', async () => {
+      renderEditor({
+        refId: '',
+        queryType: QueryType.Steps,
+        queryByResults: 'resultsQuery',
+        queryBySteps: '',
+        stepsTake: 2000,
+      } as unknown as ResultsQuery);
+      const takeInput = screen.getByPlaceholderText('Enter record count');
+
+      // User tries to enter a non-numeric value
+      await userEvent.clear(takeInput);
+      await userEvent.type(takeInput, 'abc');
+      await waitFor(() => {
+        expect(takeInput).toHaveValue(null);
+      });
+
+      // User enters a valid numeric value
+      await userEvent.clear(takeInput);
+      await userEvent.type(takeInput, '500');
+      await waitFor(() => {
+        expect(takeInput).toHaveValue(500);
+      });
+    });
+  });
 });
 
 it('should load part numbers on mount', async () => {