+++ src/datasources/test-plans/module.ts
@@ -2,7 +2,9 @@ import { DataSourcePlugin } from '@grafana/data';
 import { TestPlansDataSource } from './TestPlansDataSource';
 import { TestPlansQueryEditor } from './components/TestPlansQueryEditor';
 import { HttpConfigEditor } from 'core/components/HttpConfigEditor';
+import { TestPlansVariableQueryEditor } from './components/TestPlansVariableQueryEditor';
 
 export const plugin = new DataSourcePlugin(TestPlansDataSource)
   .setConfigEditor(HttpConfigEditor)
-  .setQueryEditor(TestPlansQueryEditor);
+  .setQueryEditor(TestPlansQueryEditor)
+  .setVariableQueryEditor(TestPlansVariableQueryEditor);